+++ person/src/com/person/core/enumeration/ActionType.java
@@ -1,0 +1,11 @@
+package com.person.core.enumeration;
+
+public enum ActionType {
+    EAT,
+    WALK,
+    LEARN,
+    DANCE,
+    CODE,
+    SING,
+    PLAY_GUITAR
+}
