+++ person/src/com/person/core/enumeration/DancerType.java
@@ -1,0 +1,7 @@
+package com.person.core.enumeration;
+
+public enum DancerType {
+    LATINO ,
+    TRADITIONAL_ARMENIAN,
+    CONTEMPORARY
+}
