+++ person/src/com/person/core/enumeration/Profession.java
@@ -1,0 +1,7 @@
+package com.person.core.enumeration;
+
+public enum Profession {
+    DANCER,
+    PROGRAMMER,
+    SINGER
+}
