+++ person/src/com/person/core/enumeration/ProgrammerType.java
@@ -1,0 +1,7 @@
+package com.person.core.enumeration;
+
+public enum ProgrammerType {
+    SOFTWARE_ENGINEER,
+    HARDWARE_ENGINEER,
+    DATA_ENGINEER
+}
