+++ person/src/com/person/core/enumeration/SingerType.java
@@ -1,0 +1,7 @@
+package com.person.core.enumeration;
+
+public enum SingerType {
+    JAZZ,
+    POP,
+    ROCK
+}
