+++ person/src/com/person/core/exceptions/NotFoundException.java
@@ -1,0 +1,8 @@
+package com.person.core.exceptions;
+
+public class NotFoundException extends ValidationException {
+    public NotFoundException(String msg)
+    {
+        super(msg);
+    }
+}
