+++ person/src/com/person/core/exceptions/ValidationException.java
@@ -1,0 +1,8 @@
+package com.person.core.exceptions;
+
+public class ValidationException extends RuntimeException {
+    public ValidationException(String msg)
+    {
+        super(msg);
+    }
+}
