+++ person/src/com/person/core/service/DancerAction.java
@@ -1,0 +1,5 @@
+package com.person.core.service;
+
+public interface DancerAction extends PersonAction {
+    void dancing();
+}
