+++ person/src/com/person/core/service/PersonAction.java
@@ -1,0 +1,10 @@
+package com.person.core.service;
+
+import com.person.core.model.BasePerson;
+
+public interface PersonAction {
+    void learn();
+    void eat();
+    void walk();
+    void act(String actionType);
+}
