+++ person/src/com/person/core/service/ProgrammerAction.java
@@ -1,0 +1,5 @@
+package com.person.core.service;
+
+public interface ProgrammerAction extends PersonAction {
+    void coding();
+}
