+++ person/src/com/person/core/service/SingerAction.java
@@ -1,0 +1,6 @@
+package com.person.core.service;
+
+public interface SingerAction extends PersonAction {
+    void singing();
+    void playingGuitar();
+}
