--- person/src/homeworks/person/core/enums/ActionType.java
@@ -1,11 +1,0 @@
-package homeworks.person.core.enums;
-
-public enum ActionType {
-    EAT,
-    WALK,
-    LEARN,
-    DANCE,
-    CODE,
-    SING,
-    PLAY_GUITAR
-}
