--- person/src/homeworks/person/core/enums/DancerType.java
@@ -1,7 +1,0 @@
-package homeworks.person.core.enums;
-
-public enum DancerType {
-    LATINO ,
-    TRADITIONAL_ARMENIAN,
-    CONTEMPORARY
-}
