--- person/src/homeworks/person/core/enums/Profession.java
@@ -1,7 +1,0 @@
-package homeworks.person.core.enums;
-
-public enum Profession {
-    DANCER,
-    PROGRAMMER,
-    SINGER
-}
