--- person/src/homeworks/person/core/enums/ProgrammerType.java
@@ -1,7 +1,0 @@
-package homeworks.person.core.enums;
-
-public enum ProgrammerType {
-    SOFTWARE_ENGINEER,
-    HARDWARE_ENGINEER,
-    DATA_ENGINEER
-}
