--- person/src/homeworks/person/core/enums/SingerType.java
@@ -1,7 +1,0 @@
-package homeworks.person.core.enums;
-
-public enum SingerType {
-    JAZZ,
-    POP,
-    ROCK
-}
