--- person/src/homeworks/person/core/service/DancerAction.java
@@ -1,5 +1,0 @@
-package homeworks.person.core.service;
-
-public interface DancerAction extends PersonAction {
-    void dancing();
-}
