--- person/src/homeworks/person/core/service/PersonAction.java
@@ -1,7 +1,0 @@
-package homeworks.person.core.service;
-
-public interface PersonAction {
-    void learn();
-    void eat();
-    void walk();
-}
