--- person/src/homeworks/person/core/service/ProgrammerAction.java
@@ -1,5 +1,0 @@
-package homeworks.person.core.service;
-
-public interface ProgrammerAction extends PersonAction {
-    void coding();
-}
