--- person/src/homeworks/person/core/service/SingerAction.java
@@ -1,6 +1,0 @@
-package homeworks.person.core.service;
-
-public interface SingerAction extends PersonAction {
-    void singing();
-    void playingGuitar();
-}
