--- person/src/homeworks/person/core/service/impl/DancerActionImpl.java
@@ -1,18 +1,0 @@
-package homeworks.person.core.service.impl;
-
-import homeworks.person.core.service.DancerAction;
-
-public class DancerActionImpl implements DancerAction {
-    @Override
-    public void learn() {}
-
-    @Override
-    public void eat() {}
-
-    @Override
-    public void walk() {}
-
-    @Override
-    public void dancing() {}
-
-}
