--- person/src/homeworks/person/core/service/impl/ProgrammerActionImpl.java
@@ -1,19 +1,0 @@
-package homeworks.person.core.service.impl;
-
-import homeworks.person.core.service.ProgrammerAction;
-
-public class ProgrammerActionImpl implements ProgrammerAction {
-
-    @Override
-    public void learn() {}
-
-    @Override
-    public void eat() {}
-
-    @Override
-    public void walk() {}
-
-    @Override
-    public void coding() {}
-
-}
