--- person/src/homeworks/person/core/service/impl/SingerActionImpl.java
@@ -1,21 +1,0 @@
-package homeworks.person.core.service.impl;
-
-import homeworks.person.core.service.SingerAction;
-
-public class SingerActionImpl implements SingerAction {
-
-    @Override
-    public void learn(){}
-
-    @Override
-    public void eat() {}
-
-    @Override
-    public void walk() {}
-
-    @Override
-    public void singing() {}
-
-    @Override
-    public void playingGuitar() {}
-}
